+++ include/visionaray/math/detail/vector3.inl
@@ -100,14 +100,14 @@ inline T const* vector<3, T>::data() const
 
 template <typename T>
 MATH_FUNC
-inline T& vector<3, T>::operator[](size_t i)
+inline T& vector<3, T>::operator[](unsigned i)
 {
     return data()[i];
 }
 
 template <typename T>
 MATH_FUNC
-inline T const& vector<3, T>::operator[](size_t i) const
+inline T const& vector<3, T>::operator[](unsigned i) const
 {
     return data()[i];
 }
